chore(test): modify test
